build warning cleanup: false-positive on unsigned comparison
The addition of (F+C+1) is incorrectly detected as a signed
value. Explicitly casting it to an unsigned value (of the same size as
F and C) helps the compiler figure it out.

Before:

```
concord-bft/bftengine/src/bftengine/ViewChangeSafetyLogic.cpp:334:19: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
       if (counter == (F + C + 1))
```

+++ bftengine/src/bftengine/ViewChangeSafetyLogic.cpp
@@ -331,7 +331,7 @@ namespace bftEngine
 					{
 						size_t& counter = fastPathCounters[fast];
 						counter++;
-						if (counter == (F + C + 1))
+						if (counter == (uint16_t)(F + C + 1))
 						{
 							selectedFastElement = fast;
 							break;
